bootstrap/requirements.txt
# No external packages currently required for directory creation
python-arango
